src/requirements.txt
jsonlines==1.2.0
datasets==1.1.3
numpy==1.14.3
omegaconf==2.0.5
tqdm==4.48.2
pandas==0.23.0
pytorch_lightning==1.5.7
nltk==3.3
matplotlib==2.2.2
networkx==2.1
torch==1.10.1
hydra_core==1.0.4
hydra==2.5
pickle5==0.0.12
ruamel.base==1.0.0
scikit_learn==1.1.1
transformers==4.20.0
